Add missing `role`s. Minor directory rename.
