+++ packages/ui/react-ui/src/components/AnchoredOverflow/AnchoredOverflow.tsx
@@ -18,7 +18,12 @@ const AnchoredOverflowRoot = forwardRef<HTMLDivElement, AnchoredOverflowRootProp
     const { tx } = useThemeContext();
     const Root = asChild ? Slot : Primitive.div;
     return (
-      <Root {...props} className={tx('anchoredOverflow.root', 'overflow-anchored', {}, classNames)} ref={forwardedRef}>
+      <Root
+        role='none'
+        {...props}
+        className={tx('anchoredOverflow.root', 'overflow-anchored', {}, classNames)}
+        ref={forwardedRef}
+      >
         {children}
       </Root>
     );
@@ -34,7 +39,12 @@ const AnchoredOverflowAnchor = forwardRef<HTMLDivElement, AnchoredOverflowAnchor
     const { tx } = useThemeContext();
     const Root = asChild ? Slot : Primitive.div;
     return (
-      <Root {...props} className={tx('anchoredOverflow.anchor', 'overflow-anchor', {}, classNames)} ref={forwardedRef}>
+      <Root
+        role='none'
+        {...props}
+        className={tx('anchoredOverflow.anchor', 'overflow-anchor', {}, classNames)}
+        ref={forwardedRef}
+      >
         {children}
       </Root>
     );
