+++ packages/ui/react-ui/src/components/index.ts
@@ -11,7 +11,7 @@ export * from './Dialogs';
 export * from './DropdownMenu';
 export * from './Input';
 export * from './Link';
-export * from './List';
+export * from './Lists';
 export * from './Main';
 export * from './Message';
 export * from './Popover';
